refactor: implement standard IDE MARK comments, JSDoc, and modularity
- Refactor `particles.js` to extract math and animation loops into a custom state engine hook and helpers.
- Refactor `og/route.js` to split font loading and SVG layout template generation into `og-helper.js`.
- Sweep files exceeding 50 lines to apply clean standard IDE MARK comments and comprehensive JSDoc metadata.
- Clean up dead/obsolete code blocks while maintaining explanatory comments.
- Ensure 100% test passability and Biome compliance.
