+++ src/app/blog/page.js
@@ -1,3 +1,4 @@
+// MARK: - Imports
 import Link from "next/link";
 import { Suspense } from "react";
 import BackButton from "@/components/buttons/back-button";
@@ -6,11 +7,20 @@ import { getBlogPosts } from "@/lib/db/blog";
 import { formatDate } from "@/lib/utils/format-date";
 import TimeToRead from "@/lib/utils/time-to-read";
 
+// MARK: - Config & Constants
 export const metadata = {
   title: "Blog",
   description: "Read my thoughts on software development, design, and more.",
 };
 
+// MARK: - Render
+
+/**
+ * BlogPage Component - renders list of available blogs sorted by publishedAt dates.
+ *
+ * @component
+ * @returns {import("react").JSX.Element}
+ */
 export default function BlogPage() {
   const allBlogs = getBlogPosts();
