+++ src/lib/db/queries/blog-post-views.js
@@ -1,5 +1,7 @@
+// MARK: - Imports
 import { ensureTable, pool } from "../db";
 
+// MARK: - Config & Constants
 const isMock = !pool;
 
 // Initialization promise to ensure all required tables exist before any DB access
@@ -8,8 +10,11 @@ const dbInitPromise = ensureTable(
   "CREATE TABLE IF NOT EXISTS blog_views (slug TEXT PRIMARY KEY, count INTEGER DEFAULT 0);",
 );
 
+// MARK: - Helper Functions
+
 /**
  * Increments the view count for a blog post and returns the updated count.
+ *
  * @param {string} slug - The blog post slug.
  * @returns {Promise<number|null>} The updated view count or null on error.
  */
@@ -41,9 +46,9 @@ export async function incrementAndGetBlogViewCount(slug) {
   }
 }
 
-// ----------- Sample functions for mutation and query ------- //
 /**
  * Increments the view count for a blog post (does not return the count).
+ *
  * @param {string} slug - The blog post slug.
  * @returns {Promise<void>}
  */
@@ -75,6 +80,7 @@ export async function incrementBlogViewCount(slug) {
 
 /**
  * Gets the view count for a blog post.
+ *
  * @param {string} slug - The blog post slug.
  * @returns {Promise<number|null>} The current view count or null if not found.
  */
